backend/requirements.txt
asgiref==3.11.1
dj-database-url==2.1.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
pillow==12.2.0
psycopg[binary]>=3.2.10,<4
requests==2.32.3
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.9.0
